+++ example/pages/rate/index.js
@@ -4,8 +4,10 @@ Page({
   data: {
     value1: 3,
     value2: 3,
-    value3: 4,
-    value4: 2
+    value3: 3,
+    value4: 2.5,
+    value5: 4,
+    value6: 3
   },
 
   onChange(event) {