+++ packages/datetime-picker/index.ts
@@ -1,13 +1,14 @@
 import { VantComponent } from '../common/component';
 import { isDef } from '../common/utils';
+import { pickerProps } from '../picker/shared';
 
 const currentYear = new Date().getFullYear();
 
-function isValidDate(date) {
+function isValidDate(date: number) {
   return isDef(date) && !isNaN(new Date(date).getTime());
 }
 
-function range(num, min, max) {
+function range(num: number, min: number, max: number) {
   return Math.min(Math.max(num, min), max);
 }
 
@@ -15,7 +16,7 @@ function padZero(val: string | number): string {
   return `00${val}`.slice(-2);
 }
 
-function times(n: number, iteratee: (number) => string): string[] {
+function times(n: number, iteratee: (index: number) => string): string[] {
   let index = -1;
   const result = Array(n);
 
@@ -37,27 +38,18 @@ function getMonthEndDay(year: number, month: number): number {
   return 32 - new Date(year, month - 1, 32).getDate();
 }
 
+const defaultFormatter = (_, value) => value;
+
 VantComponent({
+  classes: ['active-class', 'toolbar-class', 'column-class'],
+
   props: {
+    ...pickerProps,
+    formatter: {
+      type: Function,
+      value: defaultFormatter
+    },
     value: null,
-    title: String,
-    loading: Boolean,
-    itemHeight: {
-      type: Number,
-      value: 44
-    },
-    visibleItemCount: {
-      type: Number,
-      value: 5
-    },
-    confirmButtonText: {
-      type: String,
-      value: '确认'
-    },
-    cancelButtonText: {
-      type: String,
-      value: '取消'
-    },
     type: {
       type: String,
       value: 'datetime'
@@ -98,7 +90,7 @@ VantComponent({
   },
 
   watch: {
-    value(val) {
+    value(val: any) {
       const { data } = this;
       val = this.correctValue(val);
       const isEqual = val === data.innerValue;
@@ -118,7 +110,9 @@ VantComponent({
   methods: {
     getPicker() {
       if (this.picker == null) {
-        const picker = this.picker = this.selectComponent('.van-datetime-picker');
+        const picker = (this.picker = this.selectComponent(
+          '.van-datetime-picker'
+        ));
         const { setColumnValues } = picker;
         picker.setColumnValues = (...args: any) =>
           setColumnValues.apply(picker, [...args, false]);
@@ -127,11 +121,12 @@ VantComponent({
     },
 
     updateColumns() {
+      const { formatter = defaultFormatter } = this.data;
       const results = this.getRanges().map(({ type, range }, index) => {
         const values = times(range[1] - range[0] + 1, index => {
           let value = range[0] + index;
           value = type === 'year' ? `${value}` : padZero(value);
-          return value;
+          return formatter(type, value);
         });
 
         return { values };
@@ -308,23 +303,29 @@ VantComponent({
 
     updateColumnValue(value) {
       let values = [];
-      const { data } = this;
+      const { type, formatter = defaultFormatter } = this.data;
       const picker = this.getPicker();
 
-      if (data.type === 'time') {
+      if (type === 'time') {
         const pair = value.split(':');
-        values = [pair[0], pair[1]];
+        values = [
+          formatter('hour', pair[0]),
+          formatter('minute', pair[1])
+        ];
       } else {
         const date = new Date(value);
-        values = [`${date.getFullYear()}`, padZero(date.getMonth() + 1)];
-        if (data.type === 'date') {
-          values.push(padZero(date.getDate()));
+        values = [
+          formatter('year', `${date.getFullYear()}`),
+          formatter('month', padZero(date.getMonth() + 1))
+        ];
+        if (type === 'date') {
+          values.push(formatter('day', padZero(date.getDate())));
         }
-        if (data.type === 'datetime') {
+        if (type === 'datetime') {
           values.push(
-            padZero(date.getDate()),
-            padZero(date.getHours()),
-            padZero(date.getMinutes())
+            formatter('day', padZero(date.getDate())),
+            formatter('hour', padZero(date.getHours())),
+            formatter('minute', padZero(date.getMinutes()))
           );
         }
       }