+++ packages/dialog/index.ts
@@ -12,6 +12,7 @@ VantComponent({
     title: String,
     message: String,
     useSlot: Boolean,
+    className: String,
     asyncClose: Boolean,
     messageAlign: String,
     showCancelButton: Boolean,
@@ -100,7 +101,7 @@ VantComponent({
       }
       this.$emit('close', action);
 
-      //把 dialog 实例传递出去，可以通过 stopLoading() 在外部关闭按钮的 loading
+      // 把 dialog 实例传递出去，可以通过 stopLoading() 在外部关闭按钮的 loading
       this.$emit(action, { dialog: this });
 
       const callback = this.data[action === 'confirm' ? 'onConfirm' : 'onCancel'];