+++ packages/stepper/index.ts
@@ -48,7 +48,7 @@ VantComponent({
 
       const newValue = this.range(value);
 
-      if (typeof newValue === 'number' && value !== newValue) {
+      if (typeof newValue === 'number' && +this.data.value !== newValue) {
         this.set({ value: newValue });
       }
     }
@@ -77,6 +77,7 @@ VantComponent({
 
     // limit value range
     range(value) {
+      value = String(value).replace(/[^0-9.-]/g, '');
       return Math.max(Math.min(this.data.max, value), this.data.min);
     },