+++ docs/markdown/changelog.md
@@ -1,5 +1,61 @@
 ## 更新日志
 
+## [v0.5.12](https://github.com/youzan/vant-weapp/tree/v0.5.12)
+`2019-05-05`
+
+**Improvements**
+
+- Field: 新增 password 属性 [\#1583](https://github.com/youzan/vant-weapp/pull/1583)
+- Dialog: 新增 class-name 属性 [\#1601](https://github.com/youzan/vant-weapp/pull/1601)
+- SubmitBar: 新增 tip-icon 属性 [\#1593](https://github.com/youzan/vant-weapp/pull/1593)
+
+**Bug Fixes**
+
+- 修复 Stepper 不能输入小数的问题 [\#1600](https://github.com/youzan/vant-weapp/pull/1600)
+- 修复 Stepper 粘贴中文字符时报错的问题 [\#1599](https://github.com/youzan/vant-weapp/pull/1599)
+- 修复 Icon new、question 图标展示不全的问题 [\#1595](https://github.com/youzan/vant-weapp/pull/1595)
+- 修复 SubmitBar 使用标签选择器导致框架报错的问题 [\#1577](https://github.com/youzan/vant-weapp/pull/1577)
+- 修复 NoticeBar 使用标签选择器导致框架报错的问题 [\#1574](https://github.com/youzan/vant-weapp/pull/1574)
+
+
+## [v0.5.11](https://github.com/youzan/vant-weapp/tree/v0.5.11)
+`2019-04-24`
+
+**Bug Fixes**
+
+- Area: 修复同步设置areaList时不显示选项 [\#1556](https://github.com/youzan/vant-weapp/pull/1556)
+- Slider: 修复拖拽同时设置 value 时拖拽失效 [\#1553](https://github.com/youzan/vant-weapp/pull/1553)
+- Collapse: 修复 collapse-item 销毁时未从 collapse 中注销 [\#1517](https://github.com/youzan/vant-weapp/pull/1517)
+
+**Improvements**
+
+- Icon: 升级 @vant/icons 至1.1.6 [\#1560](https://github.com/youzan/vant-weapp/pull/1560)
+- DatetimePicker: 新增 formatter 属性、 新增外部样式类 [\#1558](https://github.com/youzan/vant-weapp/pull/1558)
+- Badge: 新增 click 事件 [\#1557](https://github.com/youzan/vant-weapp/pull/1557)
+- Toast: 新增 onClose 属性、 新增插槽 [\#1552](https://github.com/youzan/vant-weapp/pull/1552)
+- Picker: 新增 default-index 属性 [\#1540](https://github.com/youzan/vant-weapp/pull/1540)
+- Collapse: 新增 clickable 属性 [\#1538](https://github.com/youzan/vant-weapp/pull/1538)
+- Notify: 新增 zIndex 属性 [\#1535](https://github.com/youzan/vant-weapp/pull/1535)
+- SubmitBar: 新增 decimal-length 属性 [\#1529](https://github.com/youzan/vant-weapp/pull/1529)
+- Field: 使用 view 模拟 textarea 的 placeholder, 修复 placeholder 漂移问题 [\#1527](https://github.com/youzan/vant-weapp/pull/1527)
+- Cell: 新增 useLabelSlot 属性、 新增 label 插槽 [\#1510](https://github.com/youzan/vant-weapp/pull/1510)
+- Feild: 新增 error-message-align 属性 [\#1509](https://github.com/youzan/vant-weapp/pull/1509)
+
+## [v0.5.10](https://github.com/youzan/vant-weapp/tree/v0.5.10)
+`2019-04-11`
+
+**Bug Fixes**
+
+- Collapse: 优化性能、修复初始化未渲染时展开状态错误 [\#1506](https://github.com/youzan/vant-weapp/pull/1506)
+
+**Improvements**
+
+- SwipeCell: 优化性能、优化纵向滑动阻止页面滚动 [\#1501](https://github.com/youzan/vant-weapp/pull/1501)
+- 修复文档中 changelog 链接错误 [\#1491](https://github.com/youzan/vant-weapp/pull/1491)
+- 添加 es5 版本指引文档 [\#1489](https://github.com/youzan/vant-weapp/pull/1489)
+
+
+
 ## [v0.5.9](https://github.com/youzan/vant-weapp/tree/v0.5.9)
 `2019-04-03`